+++ ChangeLog.md
@@ -2,6 +2,7 @@
   * Add help system and initial help file
   * Make absolute damage visible
   * Add 'average' roll for blueprints
+  * Update spacing for movement summary
 
 #2.2.10
   * Fix detailed export of module reinforcement packages