+++ ChangeLog.md
@@ -3,6 +3,7 @@
   * Make absolute damage visible
   * Add 'average' roll for blueprints
   * Update spacing for movement summary
+  * Provide damage dealt statistics for both shields and hull
 
 #2.2.10
   * Fix detailed export of module reinforcement packages