+++ ChangeLog.md
@@ -6,6 +6,8 @@
   * Provide damage dealt statistics for both shields and hull
   * Damage dealt panel only shows enabled weapons
   * Add engagement range to damage received panel
+  * Handle burst rate of fire as an absolute number rather than a perentage modification
+  * Ensure that clip values are always rounded up
 
 #2.2.10
   * Fix detailed export of module reinforcement packages